refactor: reorganize Todo domain structure by moving exceptions and entities to dedicated modules, and update imports accordingly
