+++ dddpy/domain/todo/entities/__init__.py
@@ -1,0 +1,3 @@
+from .todo import Todo
+
+__all__ = ['Todo']
