+++ dddpy/domain/todo/exceptions/__init__.py
@@ -1,0 +1,13 @@
+"""Todo exceptions"""
+
+from .todo_already_completed_error import TodoAlreadyCompletedError
+from .todo_already_started_error import TodoAlreadyStartedError
+from .todo_not_found_error import TodoNotFoundError
+from .todo_not_started_error import TodoNotStartedError
+
+__all__ = [
+    'TodoNotFoundError',
+    'TodoNotStartedError',
+    'TodoAlreadyCompletedError',
+    'TodoAlreadyStartedError',
+]
