+++ dddpy/domain/todo/exceptions/todo_already_completed_error.py
@@ -1,0 +1,10 @@
+"""TodoAlreadyCompletedError exception"""
+
+
+class TodoAlreadyCompletedError(Exception):
+    """TodoAlreadyCompletedError is an error that occurs when a Todo is already completed."""
+
+    message = 'The Todo is already completed.'
+
+    def __str__(self):
+        return TodoAlreadyCompletedError.message
