+++ dddpy/domain/todo/exceptions/todo_already_started_error.py
@@ -1,0 +1,10 @@
+"""TodoAlreadyStartedError exception"""
+
+
+class TodoAlreadyStartedError(Exception):
+    """TodoAlreadyStartedError is an error that occurs when a Todo is already started."""
+
+    message = 'The Todo is already started.'
+
+    def __str__(self):
+        return TodoAlreadyStartedError.message
