+++ dddpy/domain/todo/exceptions/todo_not_found_error.py
@@ -1,0 +1,10 @@
+"""TodoNotFoundError exception"""
+
+
+class TodoNotFoundError(Exception):
+    """TodoNotFoundError is an error that occurs when a Todo is not found."""
+
+    message = 'The Todo you specified does not exist.'
+
+    def __str__(self):
+        return TodoNotFoundError.message
