+++ dddpy/domain/todo/exceptions/todo_not_started_error.py
@@ -1,0 +1,10 @@
+"""TodoNotStartedError exception"""
+
+
+class TodoNotStartedError(Exception):
+    """TodoNotStartedError is an error that occurs when a Todo is not started."""
+
+    message = 'The Todo is not started.'
+
+    def __str__(self):
+        return TodoNotStartedError.message
