+++ dddpy/domain/todo/value_objects/todo_status.py
@@ -1,0 +1,11 @@
+"""Value objects for Todo status."""
+
+from enum import Enum
+
+
+class TodoStatus(Enum):
+    """Value object representing the status of a Todo"""
+
+    NOT_STARTED = 'not_started'
+    IN_PROGRESS = 'in_progress'
+    COMPLETED = 'completed'
